+++ include/overlay006/tv_episode.h
@@ -1,5 +1,5 @@
-#ifndef POKEPLATINUM_OV6_02246444_H
-#define POKEPLATINUM_OV6_02246444_H
+#ifndef POKEPLATINUM_OV6_TV_EPISODE_H
+#define POKEPLATINUM_OV6_TV_EPISODE_H
 
 #include "struct_defs/struct_0202440C.h"
 
@@ -14,4 +14,4 @@ int TVEpisode_GetSegmentID(const TVEpisode *episode);
 TVEpisode *TVEpisode_New(FieldSystem *fieldSystem, int programType, int segmentID);
 void TVEpisode_Free(TVEpisode *episode);
 
-#endif // POKEPLATINUM_OV6_02246444_H
+#endif // POKEPLATINUM_OV6_TV_EPISODE_H
