+++ include/struct_defs/tv_segment_contest_hall_showcased_pokemon.h
