+++ include/unk_0202E2CC.h
@@ -11,7 +11,7 @@
 #include "struct_defs/struct_0202E81C.h"
 #include "struct_defs/struct_0202E828.h"
 #include "struct_defs/struct_0202E834.h"
-#include "struct_defs/tv_episode_segment_contest_hall_showcased_pokemon.h"
+#include "struct_defs/tv_segment_contest_hall_showcased_pokemon.h"
 
 int TVBroadcast_SaveSize(void);
 void TVBroadcast_Init(TVBroadcast *broadcast);
