+++ src/encounter.c
@@ -44,7 +44,7 @@
 #include "system_flags.h"
 #include "system_vars.h"
 #include "trainer_data.h"
-#include "tv_episode_segment.h"
+#include "tv_segment.h"
 #include "unk_0202F1D4.h"
 #include "unk_0203D1B8.h"
 #include "unk_020528D0.h"
