+++ src/meson.build
@@ -258,7 +258,7 @@ pokeplatinum_c = files(
     'roaming_pokemon.c',
     'npc_trade_task.c',
     'cutscenes/boat_cutscene/boat_cutscene.c',
-    'tv_episode_segment.c',
+    'tv_segment.c',
     'rankings_machine.c',
     'unk_02070428.c',
     'field_move_tasks.c',
@@ -459,7 +459,7 @@ pokeplatinum_c = files(
     'overlay006/hm_cut_in.c',
     'overlay006/roamer_after_battle.c',
     'overlay006/npc_trade.c',
-    'overlay006/ov6_02246444.c',
+    'overlay006/tv_episode.c',
     'overlay006/ov6_022465FC.c',
     'overlay006/tv_commercials.c',
     'overlay006/special_dates.c',
