+++ src/overlay104/ov104_02239130.c
@@ -26,7 +26,7 @@
 #include "pokemon.h"
 #include "savedata.h"
 #include "sound.h"
-#include "tv_episode_segment.h"
+#include "tv_segment.h"
 #include "unk_020363E8.h"
 #include "unk_02049D08.h"
 #include "unk_0209B6F8.h"
