+++ src/scrcmd_party.c
@@ -21,7 +21,7 @@
 #include "pokemon.h"
 #include "ribbon.h"
 #include "save_player.h"
-#include "tv_episode_segment.h"
+#include "tv_segment.h"
 #include "unk_02017038.h"
 #include "unk_02054884.h"
 #include "unk_0205DFC4.h"
