+++ src/underground/traps.c
@@ -53,7 +53,7 @@
 #include "system_vars.h"
 #include "terrain_collision_manager.h"
 #include "trainer_info.h"
-#include "tv_episode_segment.h"
+#include "tv_segment.h"
 #include "underground.h"
 #include "unk_020366A0.h"
 #include "vars_flags.h"
