+++ src/unk_02049D08.c
@@ -37,7 +37,7 @@
 #include "system_flags.h"
 #include "system_vars.h"
 #include "trainer_info.h"
-#include "tv_episode_segment.h"
+#include "tv_segment.h"
 #include "underground.h"
 #include "unk_0202D05C.h"
 #include "unk_0204AEE8.h"
